DXP-1310 Implement way to optionally skip executing integration tests
