+++ src/Impl/Utils/DataValidator.php
@@ -4,39 +4,24 @@ namespace Streamx\Clients\Ingestion\Impl\Utils;
 
 class DataValidator
 {
-    private object $object;
+    private PropertyRetriever $propertyRetriever;
 
     private function __construct(object $object)
     {
-        $this->object = $object;
+        $this->propertyRetriever = PropertyRetriever::for($object);
     }
 
-    public static function for(object $object): DataValidator
+    public static function for(object $object): self
     {
         return new self($object);
     }
 
     public function require(string $property)
     {
-        if (property_exists($this->object, $property)) {
-            return $this->requireNonNull($this->object->{$property}, $property);
+        $value = $this->propertyRetriever->retrieveNullable($property);
+        if ($value) {
+            return $value;
         }
         throw new DataValidationException("Property [$property] is required");
-    }
-
-    public function retrieveNullable(string $property)
-    {
-        if (property_exists($this->object, $property)) {
-            return $this->object->{$property};
-        }
-        return null;
-    }
-
-    private function requireNonNull($value, string $name)
-    {
-        if ($value == null) {
-            throw new DataValidationException("Property [$name] is required");
-        }
-        return $value;
     }
 }
